refactor: migrate method board to orbits
