+++ apps/web/menu/orbits.html
@@ -2,17 +2,28 @@
 <html lang="ko">
 <head>
   <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="refresh" content="0; url=orbits/">
   <title>ORBITS | Two.4</title>
+  <script>
+    window.location.replace('orbits/');
+  </script>
   <style>
-    body{margin:0;background:linear-gradient(180deg,#03040a,#0a0713,#120024);color:cyan;
-         font-family:system-ui,Segoe UI,Roboto,sans-serif;text-align:center;padding-top:160px}
-    .page-title{opacity:.9}
+    body {
+      margin: 0;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      min-height: 100vh;
+      font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+      background: radial-gradient(circle at top, rgba(10, 8, 24, 0.95), #05040f);
+      color: #e7ffff;
+    }
+    a {
+      color: #64d7ff;
+    }
   </style>
 </head>
 <body>
-  <div include-html="header.html"></div>
-  <h2 class="page-title">[ ORBITS (INDICATORS) ] 페이지는 현재 업데이트 진행 중입니다.</h2>
-  <script src="../include.js"></script>
+  <p>Redirecting to <a href="orbits/">ORBITS (SIGNALS / INDICATORS)</a>…</p>
 </body>
 </html>
